Fix a horrible bug in the Solved() function.
The Solve() function was comparing the Guess board with the Grid board. For cases where we had deserialized the input game, the two boards start as equal (and thus look solved) even before we run the solver.

Make Solve() check to see if the state of the Guess board is a valid solution for the game.

Update the Print() function to also print the col/row pos/neg counts for the Frames and Grid boards.