Add test for Deserialize, fix bug
Deserialize blindly took the index of several runes. If any of these were not found it would get -1 as a result and would happily try to index in with that. Always check return value to see if the expected rune was actually not found. And, add a test for this.