Start adding 'flip'
Instead of coding one set of rules to look at rows and another set of
rules to look at columns, allow for the board to be flipped along its
diagonal axis. This means we can run the row rules against the rows,
then flip the board and run the row rules again against what were the
columns.